request

+++ lib/src/main/kotlin/sp/kx/tlsmessages/TLSIssuer.kt
@@ -4,7 +4,7 @@ import java.util.UUID
 import javax.crypto.SecretKey
 
 class TLSIssuer(
-    val method: Byte,
+    val method: Int,
     val query: ByteArray,
     val key: SecretKey,
     val id: UUID,
